fix(planmode): allow safe null redirects
