+++ internal/planmode/policy_test.go
@@ -153,6 +153,63 @@ func TestDecideDeclaredReadOnlyBashCommandStillBlocksShellSyntax(t *testing.T) {
 	}
 }
 
+func TestDecideAllowsBashSafeNullRedirects(t *testing.T) {
+	bash := func(cmd string) Call {
+		args, err := json.Marshal(map[string]any{"command": cmd})
+		if err != nil {
+			t.Fatal(err)
+		}
+		return Call{Name: "bash", Args: args}
+	}
+
+	for _, cmd := range []string{
+		"git log 2>/dev/null",
+		"git log 2> /dev/null",
+		"git log >/dev/null",
+		"git log >>/dev/null",
+		"git log &>/dev/null",
+		"git log &>> /dev/null",
+		"git log 2>$null",
+		"git log 2>NUL",
+		"git log 2>&1",
+		"2>/dev/null git log",
+	} {
+		t.Run(cmd, func(t *testing.T) {
+			if decision := (Policy{}).Decide(bash(cmd)); decision.Blocked {
+				t.Fatalf("safe redirect command should be allowed in plan mode: %s", decision.Message)
+			}
+		})
+	}
+}
+
+func TestDecideBashSafeRedirectsStayConservative(t *testing.T) {
+	bash := func(cmd string) Call {
+		args, err := json.Marshal(map[string]any{"command": cmd})
+		if err != nil {
+			t.Fatal(err)
+		}
+		return Call{Name: "bash", Args: args}
+	}
+
+	for _, cmd := range []string{
+		"git log >/tmp/out",
+		"git log >/dev/nullish",
+		"git log 2>$nullish",
+		"git log 2>nul.txt",
+		"git log < /dev/null",
+		"git log 2>/dev/null && rm -rf /tmp/x",
+		"git diff --output changes.patch 2>/dev/null",
+		`git log "2>/dev/null"`,
+		`git log 2\>/dev/null`,
+	} {
+		t.Run(cmd, func(t *testing.T) {
+			if decision := (Policy{}).Decide(bash(cmd)); !decision.Blocked {
+				t.Fatal("unsafe or non-operator redirect-looking command should be blocked in plan mode")
+			}
+		})
+	}
+}
+
 func TestDecideDeclaredReadOnlyBashCommandIgnoresShellInterpreters(t *testing.T) {
 	p := Policy{ReadOnlyCommands: []string{"bash", "gh", "gh issue", "gh issue close"}}
 	args, err := json.Marshal(map[string]any{"command": "bash run-anything"})
